+++ source/client/ui.py
@@ -4,7 +4,7 @@ import plotly
 import plotly.graph_objs as go
 import numpy as np
 
-from source.server import server
+from source.back import back
 
 import source._helpers as hlp
